chore: rename fundSuggestData
